Introduced permission-restriction for delegated dataspaces. I've decided to use the 2 permission bits of Sm for that: - It's exactly what we want: bind the permissions to the   capability we pass around - It's simple: the only thing one has to do when delegating   a DS with restricted rights is to call a method of DS and   specify what permissions to give away. The implementation   is also simple. Just a few lines in ChildManager to look   at the permission bits and a method in DataSpace. - It's lightweight (compared to the approach described below). The only disadvantage I see is that is kind of a misuse of the semaphore permission bits. But I don't see any NOVA-change coming that will make this approach impossible. So, looking at the advantages, I can live with that.
The alternative approach we thought about was to introduce
something like an announcable pagefault handler for own
dataspaces. That is, you can create a dataspace and tell
your parent that you're going to handle the pagefaults. The
parent will then call a portal you gave him for every
pagefault and you can decide what to do. I think, we should
introduce this mechanism but not for the current problem.
Because IMO it's an overkill for that and to inconvenient
to use. But it will be nice for things like filesystems
with memory mapped files, ROM provider, ...

Note that I had to remove the usage of the DataSpace-semaphore
in the producer/consumer stuff. Because 1. the permission bits
might not allow that and 2. it's better to not assume that the
DataSpace cap is a semaphore anyway. So, from now on it's an
opaque cap and nobody should assume anything about it.